requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.1
Werkzeug==2.0.2
praw==7.5.0

